Add AnalysisBlinding to anita_3and4 branch.
Add AnitaBlinding namespace, which is called by AnitaEventCalibrator.
This AnitaBlinding.h and .cxx contain a list of strategies and functions to
implement them.

The AnitaEventCalibrator instance calls
AnalysisBlinding::applyBlinding(event).
This modifies the event in the desired way.

Blinding methods can be explicitly set by AnitaBlinding::setStrategy();
My inserted event blinding is currently the default, but I imagine this
will change in the near future.

Currently prints a warning message if not explicitly set or there is
no blinding when AnitaVersion::get() returns 3.