Properly init AnitaEventCalibrator arrays to zero in constructor
This is a workaround to an intermittent bug in calibrating ANITA-3
flight which I've discovered but not fixed! Occasionally there will be
a waveform to which an extra point is somehow assigned. I guess this
is a problem in the unwrapChannel function. This read from
uninitialized memory in AnitaEventCalibrator, which could get a sample
so huge it ultimately crashed the more fragile functionality
downstream in the analysis (e.g. sine subtraction). While this bug
should be fixed, mucking with the low level unwrapping at this stage
of the analysis is not smart. An example event is ANITA-3, run 389,
eventNumber 71405902. To see this in action, edit the intialization
from zero to a crazy recognizable number. Some number of channels with
have 253 points, with the last being the crazy number. This commit is
just a work around, tantamount to adding a near-zero noise point.